Fast Fellowship Promote Tracks (#356)
Add fast promotion tracks with 7 days (instead of 30 days) voting
period.
The tracks have a 50% turnout and 66% approval criteria. To be used in
combination with https://github.com/paritytech/polkadot-sdk/pull/4778.

It could be that the 50% turnout is too much for tracks I and II, but
for III there are not so many fellows so we can expect them to all vote
IMHO.

---------

Signed-off-by: Oliver Tale-Yazdi <oliver.tale-yazdi@parity.io>
Co-authored-by: joe petrowski <25483142+joepetrowski@users.noreply.github.com>