Remove pov-reclaim from kusama system-chains (#423)
I propose to move PoV-reclaim to the next release.

It can happen that the amount to reclaim is overestimated
https://github.com/paritytech/polkadot-sdk/issues/5229.

Improvements (https://github.com/paritytech/polkadot-sdk/pull/5281,
https://github.com/paritytech/polkadot-sdk/pull/5273) and a proper fix
(https://github.com/paritytech/polkadot-sdk/pull/5234) are in the works,
but should be tested again.

I expect the chances that the system-chains run into this scenario to be
low, but it does not look like activity is so high that the inclusion of
reclaim is urgent.

---------

Co-authored-by: Bastian Köcher <git@kchr.de>