Add coretime chainspecs for kusama and polkadot and presets for live chains (#432)
Add the invulnerables set directly to the preset and generate
chain-specs for the "live" (production) networks directly from the
`chain-spec-builder` instead of using a bash script to post-process
these from generated local chain-specs.

The Kusama Coretime Chain is added here for completeness and a sanity
check with the original genesis chain-specs.

Note that the SS58 prefix is ignored when generating the chain-specs, as
noted in https://github.com/paritytech/polkadot-sdk/issues/5271

Supersedes https://github.com/seadanda/runtimes/pull/4 since it seems
you can't change the target branch of a PR across forks.

Note that this branch was made from the tip of another, so the commit
history includes changes from that branch.

---------

Co-authored-by: Bastian Köcher <git@kchr.de>
Co-authored-by: Branislav Kontur <bkontur@gmail.com>
Co-authored-by: Tom <tsenseless@gmail.com>
Co-authored-by: fellowship-merge-bot[bot] <151052383+fellowship-merge-bot[bot]@users.noreply.github.com>