Add "Spokesperson" proxy type to Kusama (#430)
A "Spokesperson" is a proxy type only allowed to make remarks in the
name of the proxied account. Useful for DAOs with an account in the
relay chain that need to make a statement. This will also allow the
representative to act on behalf of the DAO in websites like
subsquare/polkassembly to edit proposals for example.

Co-authored-by: Bastian Köcher <git@kchr.de>
Co-authored-by: fellowship-merge-bot[bot] <151052383+fellowship-merge-bot[bot]@users.noreply.github.com>