Coretime Chain: mitigate behaviour with many assignments on one core (#434)
We can handle a maximum of 28 assignments inside one XCM, while it's
possible to have 80 (if a region is interlaced 79 times).
This can be chunked on the coretime chain side but currently the
scheduler on the relay makes assumptions that means we can't send more
than one chunk for a given core.

I've made a mitigation here that requires no sdk release, but it just
truncates the additional assignments until we can extend the relay to
support this. This means that the first 27 assignments are taken, the
final 28th is used to pad with idle to complete the mask (the relay also
assumes that every schedule is complete). Any other assignments are
dropped.

---------

Co-authored-by: Bastian Köcher <git@kchr.de>
Co-authored-by: fellowship-merge-bot[bot] <151052383+fellowship-merge-bot[bot]@users.noreply.github.com>