Proper handling for legacy parachain leases with gaps in coretime migration (#426)
While testing https://github.com/polkadot-fellows/runtimes/pull/401 I
noticed that there are two parachains with gaps in their leases which
are not migrated to coretime correctly. These are para ids 3359 and
3388.

The problem is that the migration obtains the list of parachain ids
which should be migrated from `paras::Parachains`
([here](https://github.com/paritytech/polkadot-sdk/blob/1f49358db0033e57a790eac6daccc45beba81863/polkadot/runtime/parachains/src/coretime/migration.rs#L114))
and there are no entries for para ids which are not active at the
moment.

Paras 3359 and 3388 are not active in the current lease period (excerpt
from `slots::leases()`:
```
[
      3,359
    ]
    [
      null
      [
        14xQXJdUDC1pzyt8y3z27ANiUBgP7zTSaYutaLELJoyQrdLP
        2,010,000,000,000
      ]
      [
        14xQXJdUDC1pzyt8y3z27ANiUBgP7zTSaYutaLELJoyQrdLP
        2,010,000,000,000
      ]
      <skipped>
```
and:
```
[
      3,388
    ]
    [
      null
      [
        1TpMimWf8NvrusAQ36hVdgMpYhhtojg5Nw41CuVxG2zHPDg
        31,000,000,000,000
      ]
      [
        1TpMimWf8NvrusAQ36hVdgMpYhhtojg5Nw41CuVxG2zHPDg
        31,000,000,000,000
      ]
      <skipped>
```
And they have got no entries in `paras::parachains()`:

```
[ 1,000 1,001 1,002 1,004 2,000 2,002 2,004 2,006 2,008 2,012 2,013 2,025 2,026 2,030 2,031 2,032 2,034 2,035 2,037 2,040 2,043 2,046 2,048 2,051 2,053 2,056 2,086 2,090 2,091 2,092 2,093 2,094 2,101 2,104 2,106 3,333 3,338 3,340 3,344 3,345 3,346 3,354 3,358 3,366 3,367 3,369 3,370 3,373 3,375 3,377 ]
```

As a result the migration skips them which is wrong.

A proper fix for this issue will require a new polkadot-sdk release and
a bump in the runtimes repo which requires time and effort. To avoid
this the PR moves the coretime migration from polkadot-sdk to the
fellowship repo.

---------

Co-authored-by: Bastian Köcher <git@kchr.de>
Co-authored-by: Dónal Murray <donalm@seadanda.dev>
Co-authored-by: fellowship-merge-bot[bot] <151052383+fellowship-merge-bot[bot]@users.noreply.github.com>