Kusama: Make the current inflation formula adjustable.  (#364)
This PR does not alter the inflation of Kusama, but instead make the
current parameters of the system fully configurable.

## Parameters

The parameters are as follows: 

```rust
/// Minimum inflation rate used to calculate era payouts.
pub static MinInflation: Perquintill = Perquintill::from_rational(25u64, 1000);

/// Maximum inflation rate used to calculate era payouts.
pub static MaxInflation: Perquintill = Perquintill::from_percent(10);

/// Ideal stake ratio used to calculate era payouts.
pub static IdealStake: Perquintill = Perquintill::from_percent(75);

/// Falloff used to calculate era payouts.
pub static Falloff: Perquintill = Perquintill::from_percent(5);

/// Whether to use auction slots or not in the calculation of era payouts, then we subtract
/// `num_auctioned_slots.min(60) / 200` from `ideal_stake`.
///
/// That is, we assume up to 60 parachains that are leased can reduce the ideal stake by a
/// maximum of 30%.
///
/// With the move to agile-coretime, this parameter does not make much sense and should
/// generally be set to false.
pub static UseAuctionSlots: bool = true;
```

All of the above are exactly the current values in Kusama, and leave
everything unchanged, leading to roughly `1kKSM` minted for validators,
and around `90KSM` for treasury.

All of the above can be changed via the `Root` track only. 

Given that it is hard to come to consensus, I highly advise this PR (and
a similar PR to Polkadot) to NOT alter any parameter and leave that to
the token holders.

Just to provide one example, if we set `UseAuctionSlots = false` and
leave everything else un-changed, the amount minted per era would be:

* `807KSM` for staking
* `238KSM` for treasury

> This, in my opinion, is the most non-controversial change as the
incorporation of auctions in the inflation rate is already meaningless
with agile-coretime.

## Consideration for UIs 

This PR is shipped with a new runtime API that is only added to the
Kusama runtime: `Inflation_experimental_inflation_info`. I hope this API
is used by the UIs to show the inflation parameters of the network,
rather than re-creating the Rust logic in the front-end.

---------

Signed-off-by: Oliver Tale-Yazdi <oliver.tale-yazdi@parity.io>
Co-authored-by: Oliver Tale-Yazdi <oliver.tale-yazdi@parity.io>
Co-authored-by: s0me0ne-unkn0wn <48632512+s0me0ne-unkn0wn@users.noreply.github.com>
Co-authored-by: clangenb <37865735+clangenb@users.noreply.github.com>
Co-authored-by: Dónal Murray <donal.murray@parity.io>
Co-authored-by: Gonçalo Pestana <g6pestana@gmail.com>
Co-authored-by: joe petrowski <25483142+joepetrowski@users.noreply.github.com>
Co-authored-by: Dónal Murray <donalm@seadanda.dev>
Co-authored-by: fellowship-merge-bot[bot] <151052383+fellowship-merge-bot[bot]@users.noreply.github.com>