Fix whitelisted keys in benchmarks (#525)
Currently whitelisted keys are manually set in `dispatch_benchmark`.
Actually pallet already declare the whitelisted keys, and it is missing
many whitelist.

in asset hub the key `relay_well_known_keys::ACTIVE_CONFIG` is no longer
whitelisted. It is actually only used in the relay chain pallet
`configuration`, and this pallet would whitelist the key if it was used.
But asset hub doesn't use this pallet, so it doesn't need to whitelist
the key.

This PR will improve the benchmark precision, lower the cost of some
calls.

---------

Co-authored-by: Bastian Köcher <git@kchr.de>