Relay Chain Accounts to Asset Hub Migration (#515)
This PR is not intended to be merged into the master branch but into a
dedicated Asset Hub migration working branch. Please create such a
branch if you have the necessary permissions and change the base for
this PR.

The PR introduces two pallets designed for the Asset Hub migration,
Relay Chain migrator (rc-migrator) and Asset Hub migrator (ah-migrator).
Please refer to the code base for more docs.

Currently, these pallets handle only the migration of Relay Chain
accounts. They are intended to support the migration of all migrating
pallets and act as managers for the overall migration process.

It might be helpful to merge this PR into the dev branch to use this
initial setup as a base for further development.

[x] Does not require a CHANGELOG entry

---------

Signed-off-by: Oliver Tale-Yazdi <oliver.tale-yazdi@parity.io>
Co-authored-by: Oliver Tale-Yazdi <oliver.tale-yazdi@parity.io>