force unstake the last remaining corrupt ledger (#538)
One last staking ledger that due to the bug found last year is in an
inconsistent state.

The Support team is in touch with the owner, and force_staking them is
the preferred way.

Note that this is only an action that we are doing because the root
cause of the inconsistent stake was buggy code in the polkadot runtime,
not any user interaction.

See:
https://forum.polkadot.network/t/recover-corrupted-staking-ledgers-in-polkadot-and-kusama/9796

---------

Co-authored-by: Bastian Köcher <git@kchr.de>