Update rust and use stable for benches (#574)
<!-- Remember that you can run `/merge` to enable auto-merge in the PR
-->

<!-- Remember to modify the changelog. If you don't need to modify it,
you can check the following box.
Instead, if you have already modified it, simply delete the following
line. -->

- [x] Does not require a CHANGELOG entry

We are trying to get the command bot working. However, right now it is
timing out when trying to build the runtimes plus the omni bencher after
6 hours. This should not take 6 hours ;) So, one guess is that maybe the
nightly is doing something that takes quite long to build. We want to
try the stable compiler and see if that fixes the issue.

---------

Co-authored-by: GitHub Action <action@github.com>