[AHM] Set up AH migration testing framework for single pallets (#597)
Set up a leaner migration testing framework for single pallets and
implementing it for pallet `preimage`.

This PR introduces the traits `RcMigrationChecks` and
`AhMigrationChecks` for checks before and after a single pallet's
migration. Upon implementing these traits for a pallet, you can use
`rc_migrate` and `ah_migrate` functions to run a test on the correctness
of the pallet's migration.