[AHM] Fix accounts migration failures (#628)
The PR addresses the following issues in account migrations:

Problem 1:
An account has a free balance below the existential deposit (ED) but a
reserved balance above ED. This causes the free balance to be burned
when releasing the hold, additionally leading to a panic in the test
environment due to an existing debug_assert.

Solution:
The hold is partially released first to meet the ED requirement before
releasing the remaining hold.

Example:
https://polkadot.subscan.io/account/16HwqWaypD6acNeUu3faVGVowG86UvG1ieLrXZSWVxsaNyLS

Problem 2:
An account has no free balance but a held balance above ED. This results
in an error when attempting to recover the hold on Asset Hub (AH).

Solution:
A provider reference is added on AH if the account lacks a free balance
to meet the ED requirement.

Example:
https://polkadot.subscan.io/account/167NvqSJt5wKFDR4RFYkXA1piJfEKVH8MG4fs4nX9MnbNPQa

Problem 3:
An account’s total balance (free + reserved) is below ED, causing most
fungible implementation operations to fail.

Solution:
Such accounts are ignored and not migrated.

Example (currently only one account):
https://polkadot.subscan.io/account/15QBmCuy3aRxn8xuzfaRvzaLdawsGbPmttbcvBU8q71kEb7a