[AHM] Pallet multisig tests (#666)
1. Removing migrated multisigs from the relay chain
2. Add pre and post migration tests.
3. Check that you can re-create a Relay Chain multisig on Asset Hub and
the ID stays the same

Notice that multisigs storage items are not actually migrated to Asset
Hub, but only the multisig accounts (with non-zero balance) are, so that
they can be re-created on asset hub by users if they want to. Thus, we
only need to check that multisig storage items are deleted from the
relay chain and relevant multisig accounts are migrated to asset hub.

Finally, a separate test creates a new multisig on the relay chain with
non-zero balance and checks that the same multisig can be re-created on
asset hub while preserving its balance

---------

Signed-off-by: Oliver Tale-Yazdi <oliver.tale-yazdi@parity.io>
Co-authored-by: Oliver Tale-Yazdi <oliver.tale-yazdi@parity.io>