[AHM] Dependency fixes for integration tests and the Polkadot runtime (#751)
Dependency fixes for integration tests and the Polkadot runtime

As seen in this PR
https://github.com/polkadot-fellows/runtimes/pull/742, integration tests
are failing because some dependencies are not enabling the std feature.
That PR also shows `chain-spec-generator` failing to build its tests,
but since there have been updates to its `Cargo.toml` on master, I
suggest we wait for those changes to be merged.

The `ahm-test-westend` CI test disabled.

Additionally, this removes the `pallet-ah-migrator` dependency from the
polkadot-runtime crate.

---------

Signed-off-by: Oliver Tale-Yazdi <oliver.tale-yazdi@parity.io>
Co-authored-by: Oliver Tale-Yazdi <oliver.tale-yazdi@parity.io>