Add balance to StakingPot accounts for benchmarks for all system parachains (#729)
An
[issue](https://github.com/polkadot-fellows/runtimes/pull/709#issuecomment-2853988376)
was uncovered with the new `pay_fees` XCM benchmark. The problem arises
when the `Trader` deposits an amount into the `StakingPot`. In some
cases, if the amount is below the existential deposit (ED), the deposit
fails and is ignored. This causes the benchmark to miss recording an
expected storage write.

This PR sets an initial balance for the `StakingPot` accounts in
genesis—**only for benchmarking purposes**.


- [X] Does not require a CHANGELOG entry

---------

Signed-off-by: Oliver Tale-Yazdi <oliver.tale-yazdi@parity.io>
Co-authored-by: GitHub Action <action@github.com>
Co-authored-by: Oliver Tale-Yazdi <oliver.tale-yazdi@parity.io>