[AHM] Disable #[pallet::hooks] for migrating pallets during migration (#784)
Disable #[pallet::hooks] for migrating pallets during migration

Since there is no built-in feature to disable specific hooks at the
runtime level, we configured certain parameters to cause those hooks to
return early. To properly review this PR, you’ll need to look into the
implementation of the affected hooks.

An alternative would be to write a custom container instead of using the
AllPalletsWithSystem type for executables, but I believe that approach
would be significantly more complex and the presented solution is more
reasonable.

These disables hooks for `treasury`, `scheduler` and `fast_unstake`
pallets on RC and AH.