[AHM] Introducing the Cool-Off stage (#786)
Introducing the Cool-Off stage

Introduces the `CoolOff` stage with a `cool_off_end` block number. The
migration should be scheduled with two time points: `start` and
`cool_off_end`. When the chain reaches the `start` block, migration
begins on the Relay Chain by filtering the listed dispatchables for
extrinsics, prioritizing messages from the Asset Hub (AH), and sending a
message to AH indicating the start of the migration. Upon receiving this
message, AH also begins filtering the listed dispatchables and
prioritizes messages from the Relay Chain. The Relay Chain does not send
any data to AH until the `cool_off_end` block has passed. This cool-off
period allows time to ensure everything is functioning correctly and
gives an opportunity to halt the migration if issues arise.

---------

Signed-off-by: Oliver Tale-Yazdi <oliver.tale-yazdi@parity.io>
Co-authored-by: Oliver Tale-Yazdi <oliver.tale-yazdi@parity.io>