[AHM] Migration idempotency (#808)
Migration idempotency

This PR introduces support for running the migration multiple times from
any stage within the range `[AccountsMigrationOngoing { last_key: None
}, MigrationDone]`.

Changes:
- default values in StorageValues are excluded from migration data.
- redundant data transfers and actions that could occur during repeated
runs are filtered or disabled to minimize unnecessary logs, events and
message confirmaitons.
- proxy announcements are removed after being migrated from RC.
- treasury::LastSpendPeriod storage value is migrated.
- no checking balance correction on subsequent runs.
- added tests to verify migration idempotency.

Subsequent runs should not start from `AccountsMigrationInit` or call
`start_migration` on Asset Hub, as these step records the total
issuances and captures the checking account balances used for the
checking account correction on Asset Hub as the last step of the
migration. This part is not currently idempotent.

While it could be made idempotent - for example, by moving the checking
balance correction to a separate migration stage - such a change would
require rewriting the correction logic and its tests. Given the limited
impact, current priorities, and the fact that the balance correction
likely needs a full redesign anyway, this is not considered necessary at
the moment.

---------

Co-authored-by: Oliver Tale-Yazdi <oliver.tale-yazdi@parity.io>