[AHM] Complete account balances tests (#674)
**Migration Changes**
- If a partially migrated account has not enough free balance (_free_)
to cover for the Relay Chain existential deposit + Asset Hub existential
deposit (1 DOT + 0,01 DOT for Polkadot), we unreserve some of the
reserved balance kept on the Relay Chain to cover for the existential
deposits. (On Polkadot, if _free_ < 1,01 DOT, then we unreserve _free_ -
1,01 DOTs from the reserved balance kept on the relay chain).

**New tests**

- [x] Extensive Relay Chain pre-migration on post-migration checks on
all the accounts.
- [x] Balances cross-checks on RC and AH post-migration for checking
account balance and total issuance
- [x] AH checks on correctly migrated freezes/locks and holds/reserves


For more details, you can take a look at the comments or at [this
document](https://hackmd.io/DieSAxrLSuGgFzpgzNd24Q?both).

---------

Co-authored-by: Oliver Tale-Yazdi <oliver.tale-yazdi@parity.io>