[AHM] Optional migration manager account id (#818)
Optional migration manager account id

Adds the possibility to set a specific account id as a manager for the
migrators pallets. The manager has the similar to migrators
`AdminOrigin` (`ManagerOrigin` before) privileges except that it can not
set the manager account id via `set_manager` call. The account id later
can be an off-chain multisig set by Governance or Fellowship to manage
the migration process.