[AHM] Staking migration (#821)
Changes:
- Remove `ahm-staking-migration` feature since its not the default
- Migrate `pallet-staking` to `pallet-staking-async`

Translation works by defining a portable type. The RC types are
converted to this portable type, sent over to the AssetHub and then
translated to AssetHub types there.
The portable types do not have generics. This makes it easy to translate
them because they do not need a T Config or anything.

TODO:
- [ ] Extend tests
- [ ] Benchmarks

---------

Signed-off-by: Oliver Tale-Yazdi <oliver.tale-yazdi@parity.io>