Edit README (Release guidelines) and PR template (#745)
<!-- Remember that you can run `/merge` to enable auto-merge in the PR
-->

<!-- Remember to modify the changelog. If you don't need to modify it,
you can check the following box.
Instead, if you have already modified it, simply delete the following
line. -->

- [X] Does not require a CHANGELOG entry

This PR makes it a mandatory step to review tooling used in the Fellows
CI pipeline, as requested in
[#743](https://github.com/polkadot-fellows/runtimes/issues/743#issuecomment-2923764451).
It also removes Santi Balaguer from the process, as he is no longer
involved in releases-related comms.

---------

Co-authored-by: Bastian Köcher <git@kchr.de>