[AHM] Record migration timings (#835)
Record migration timings.

Closes: https://github.com/paritytech/polkadot-sdk/issues/9292

This PR records the start and end times of the migration in dedicated
storage items. The start is defined as the moment when the extrinsics of
the migrating pallets are locked; the end is when those locks are
released.

The PR also updates the trigger for the `AssetHubMigrationStarted` event
- it is now emitted upon transitioning to the `WaitingForAh` state
instead of `Starting`. This change ensures the event accurately reflects
the point at which the extrinsics become locked.

Co-authored-by: Oliver Tale-Yazdi <oliver.tale-yazdi@parity.io>