[AHM] Filter staking calls before migration starts (#847)
We currently only have call filter for the stages during and after them
migration - not before. This MR adds a third entry to the filter tuple
that can check in the stage before the migration starts.

Staking is disabled in this stage and a few other pallets. Indices
pallet was also forbidden on AH now to avoid insert conflicts.

---------

Signed-off-by: Oliver Tale-Yazdi <oliver.tale-yazdi@parity.io>