[AHM] Call filter configuration (#845)
The call filters configuration.

We have two migration-specific call filters: one on the Relay Chain (RC)
and another on the Asset Hub (AH).
- RC filter: configured as a tuple where the first value (ON or OFF)
indicates whether an extrinsic (or all extrinsics in a pallet) should be
filtered/locked during migration, and the second value indicates whether
it should remain locked after migration.
- AH filter: the ON/OFF configuration indicates whether an
extrinsic/pallet should be locked during migration.

The general rule is to lock any extrinsic that reads from or writes to
migrating pallets during migration, with the exception of the Balances
pallet, which is always consistent and may be altered during migration.
Migrating pallets on RC will remain locked after migration as well.

When deciding whether to lock a call, we should keep in mind that
certain service calls (e.g., clear queue) may be useful during migration
to resolve unexpected issues.

I’d like to invite you to review the entire call filter configuration
and approve this PR if you agree with it:
@ggwpez @kianenigma @bkontur @Ank4n @seadanda @acatangiu @ordian
@Overkillus @tdimitrov

---------

Signed-off-by: Oliver Tale-Yazdi <oliver.tale-yazdi@parity.io>
Co-authored-by: Oliver Tale-Yazdi <oliver.tale-yazdi@parity.io>